[RAD-1237] Added support for template.type parameter
