+++ .gitignore
@@ -27,3 +27,4 @@ build/
 bin/
 classes/
 generated/
+out/
