--- src/test/resources/sample/src/main/resources/templates/__projectName__-model/src/main/java/__packagePath__/model/ExampleTO.java
@@ -1,21 +1,0 @@
-package @packageName@.model;
-
-public class ExampleTO {
-
-  private String name;
-
-  public ExampleTO() {
-  }
-
-  public ExampleTO(String name) {
-    this.name = name;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-}
