+++ src/test/resources/sample/type1/.nontemplates
