+++ src/test/resources/sample/type1/build.gradle
@@ -1,3 +1,5 @@
+package sample.type1
+
 buildscript {
   repositories {
     mavenLocal()
