+++ src/test/resources/sample/type1/gradle.properties
