+++ src/test/resources/sample/type1/gradle/wrapper/gradle-wrapper.jar
