+++ src/test/resources/sample/type1/settings.gradle
@@ -1,3 +1,4 @@
+package sample.type1
 /*
  * This settings file was auto generated by the Gradle buildInit task
  * by 'hao' at '3/1/16 3:18 PM' with Gradle 2.8
