+++ src/test/resources/sample/type1/src/main/resources/templates/__projectName__-model/src/main/java/__packagePath__/model/ExampleOptional.java__IF_isSet__
