+++ src/test/resources/sample/type1/src/main/resources/templates/__projectName__-web/src/main/java/__packagePath__/resource/ExampleController.java
