+++ src/test/resources/sample/type1/src/main/resources/templates/__projectName__-web/src/main/resources/app-dev.properties
@@ -1,0 +1,10 @@
+##################################################################################################################
+# PROD ENVIRONMENT PROPERTIES
+#
+# These properties will override properties in config.properties, if system 
+# property configEnvironment is set to "prod".  For example, using JVM start up param:  -DconfigEnvironment=priod
+# 
+# Environment specific properties can be set for any environment by setting the above property and placing
+# a property file named config-propVal.properties where propVal is the value set for configEnvironment.
+##################################################################################################################
+
