+++ src/test/resources/sample/type1/src/main/resources/templates/__projectName__-web/src/main/resources/buildinfo.properties
