+++ src/test/resources/sample/type1/src/main/resources/templates/__projectName__-web/src/test/java/__packagePath__/service/ExampleServiceGetApiTest.java
