+++ src/test/resources/sample/type1/src/main/resources/templates/gradlew
