+++ src/test/resources/sample/type1/src/main/resources/templates/settings.gradle
@@ -1,3 +1,5 @@
+package sample.type1.src.main.resources.templates
+
 include "@projectName@-model", "@projectName@-web"
 
 // @packagePath@
