+++ src/test/resources/sample/type2/test.text
@@ -1,0 +1,1 @@
+This template type is @testBinding@
